requirements.txt
APScheduler==3.6.3
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
fire==0.2.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
pandas-datareader==0.8.1
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
supervisor==4.1.0
termcolor==1.1.0
tornado==6.0.3
tushare==1.2.50
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
wrapt==1.11.2
